fix: support default download callback option
