+++ packages/browsers/src/main.ts
@@ -25,6 +25,7 @@ export type {
 } from './install.js';
 export {
   install,
+  makeProgressCallback,
   getInstalledBrowsers,
   canDownload,
   uninstall,
@@ -40,6 +41,6 @@ export {
   createProfile,
   getVersionComparator,
 } from './browser-data/browser-data.js';
-export {CLI, makeProgressCallback} from './CLI.js';
+export {CLI} from './CLI.js';
 export {Cache, InstalledBrowser} from './Cache.js';
 export {BrowserTag} from './browser-data/types.js';
